Added initial scripts
simple_goto and vehicle_state are examples from DroneKit. mission_basic
is also based on an example with the same name, but has been altered in
order to actually fly in a closed square instead of skipping waypoints.
It can still go to the next waypoint if it the distance to the waypoint
becomes small enough. The syntax was also reformatted.