mission_basic: Directional sensor
Made the sensor simulator take into account that it can only sense the
object if the vehicle is pointing toward it (at least for now this is
the angle it should compare to). Calculate the angle and normalize it so
that it can be used to compare.

Also added some initial stuff to make polygon objects based on
points/altitudes work (although the ray casting in not immediately
necessary in this form), with multiple objects support and all.

Apparently the default copter is actually moving quite fast, so 1s
intervals of checking sensors is too high. (Note that the RTL command is
also acknowledged late, so the vehicle flies through the object anyway.)
Silence distance info when far away to keep logs readable.